(hel-671) Extraction de la popup des favoris dans un composant
